ocr-flask-api/requirements.txt
Flask==2.2.5
pillow==9.5.0
pytesseract==0.3.10